Fix punctuation and ortography
Both in English and in Polish it is a typographical error to place a full space before the question mark or the exclamation mark (or before any other punctuation, in general). The space should go after the mark.

Fixed some prepositions and made some of the sentences less ambiguous and easier to read.